requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.0
PyJWT>=2.8.0
